migrate tests and expand for produce-message setting
