+++ src/preferences/listener.test.ts
@@ -1,7 +1,14 @@
 import * as assert from "assert";
 import sinon from "sinon";
 import { ConfigurationChangeEvent, workspace } from "vscode";
-import { SSL_PEM_PATHS, SSL_VERIFY_SERVER_CERT_DISABLED } from "./constants";
+import { getExtensionContext } from "../../tests/unit/testUtils";
+import * as contextValues from "../context/values";
+import {
+  ENABLE_DIRECT_CONNECTIONS,
+  ENABLE_PRODUCE_MESSAGES,
+  SSL_PEM_PATHS,
+  SSL_VERIFY_SERVER_CERT_DISABLED,
+} from "./constants";
 import { createConfigChangeListener } from "./listener";
 import * as updates from "./updates";
 
@@ -9,12 +16,20 @@ describe("preferences/listener", function () {
   let sandbox: sinon.SinonSandbox;
   let getConfigurationStub: sinon.SinonStub;
   let onDidChangeConfigurationStub: sinon.SinonStub;
+  let setContextValueStub: sinon.SinonStub;
+
+  before(async () => {
+    // ResourceViewProvider interactions require the extension context to be set (used during changes
+    // in the direct connection preview setting)
+    await getExtensionContext();
+  });
 
   beforeEach(function () {
     sandbox = sinon.createSandbox();
     // stub the WorkspaceConfiguration and onDidChangeConfiguration emitter
     getConfigurationStub = sandbox.stub(workspace, "getConfiguration");
     onDidChangeConfigurationStub = sandbox.stub(workspace, "onDidChangeConfiguration");
+    setContextValueStub = sandbox.stub(contextValues, "setContextValue");
   });
 
   afterEach(function () {
@@ -70,4 +85,27 @@ describe("preferences/listener", function () {
 
     assert.ok(updatePreferencesStub.notCalled);
   });
+
+  for (const [previewSetting, previewContextValue] of [
+    [ENABLE_DIRECT_CONNECTIONS, contextValues.ContextValues.directConnectionsEnabled],
+    [ENABLE_PRODUCE_MESSAGES, contextValues.ContextValues.produceMessagesEnabled],
+  ]) {
+    for (const enabled of [true, false]) {
+      it(`should update the "${previewContextValue}" context value when the "${previewSetting}" setting is changed to ${enabled} (REMOVE ONCE PREVIEW SETTING IS NO LONGER USED)`, async () => {
+        getConfigurationStub.returns({
+          get: sandbox.stub().withArgs(previewSetting).returns(enabled),
+        });
+        const mockEvent = {
+          affectsConfiguration: (config: string) => config === previewSetting,
+        } as ConfigurationChangeEvent;
+        onDidChangeConfigurationStub.yields(mockEvent);
+
+        createConfigChangeListener();
+        // simulate the setting being changed by the user
+        await onDidChangeConfigurationStub.firstCall.args[0](mockEvent);
+
+        assert.ok(setContextValueStub.calledWith(previewContextValue, enabled));
+      });
+    }
+  }
 });
